add work and school to mobile version
